Introduce `hf-fail-server`
Based on what was added for cosmo, use similar logic in gimlet
to use an idol server that always returns an error code if we
can't read the chip ID.
