+++ drv/cosmo-hf/Cargo.toml
@@ -10,6 +10,7 @@ drv-hf-api = { path = "../hf-api" }
 drv-spartan7-loader-api = { path = "../spartan7-loader-api" }
 ringbuf = { path = "../../lib/ringbuf" }
 userlib = { path = "../../sys/userlib", features = ["panic-messages"] }
+drv-hf-fail-server = { path = "../hf-fail-server" }
 
 cortex-m = { workspace = true }
 idol-runtime = { workspace = true }
