+++ drv/cosmo-hf/src/hf.rs
@@ -735,216 +735,6 @@ impl NotificationHandler for ServerImpl {
     }
 }
 
-/// Dummy server which returns an error for every operation
-pub struct FailServer {
-    pub err: HfError,
-}
-
-impl idl::InOrderHostFlashImpl for FailServer {
-    fn read_id(
-        &mut self,
-        _: &RecvMessage,
-    ) -> Result<[u8; 20], RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn capacity(
-        &mut self,
-        _: &RecvMessage,
-    ) -> Result<usize, RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    /// Reads the STATUS_1 register from the SPI flash
-    fn read_status(
-        &mut self,
-        _: &RecvMessage,
-    ) -> Result<u8, RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn bulk_erase(
-        &mut self,
-        _: &RecvMessage,
-        _protect: HfProtectMode,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn page_program(
-        &mut self,
-        _: &RecvMessage,
-        _addr: u32,
-        _protect: HfProtectMode,
-        _data: LenLimit<Leased<R, [u8]>, PAGE_SIZE_BYTES>,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn page_program_dev(
-        &mut self,
-        _: &RecvMessage,
-        _: HfDevSelect,
-        _addr: u32,
-        _protect: HfProtectMode,
-        _data: LenLimit<Leased<R, [u8]>, PAGE_SIZE_BYTES>,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn read(
-        &mut self,
-        _: &RecvMessage,
-        _offset: u32,
-        _dest: LenLimit<Leased<W, [u8]>, PAGE_SIZE_BYTES>,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn read_dev(
-        &mut self,
-        _: &RecvMessage,
-        _: HfDevSelect,
-        _offset: u32,
-        _dest: LenLimit<Leased<W, [u8]>, PAGE_SIZE_BYTES>,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn sector_erase(
-        &mut self,
-        _: &RecvMessage,
-        _addr: u32,
-        _protect: HfProtectMode,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn sector_erase_dev(
-        &mut self,
-        _: &RecvMessage,
-        _: HfDevSelect,
-        _addr: u32,
-        _protect: HfProtectMode,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn bonus_page_program(
-        &mut self,
-        _: &RecvMessage,
-        _addr: u32,
-        _data: LenLimit<Leased<R, [u8]>, PAGE_SIZE_BYTES>,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn bonus_read(
-        &mut self,
-        _: &RecvMessage,
-        _offset: u32,
-        _dest: LenLimit<Leased<W, [u8]>, PAGE_SIZE_BYTES>,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn bonus_sector_erase(
-        &mut self,
-        _: &RecvMessage,
-        _addr: u32,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn get_mux(
-        &mut self,
-        _: &RecvMessage,
-    ) -> Result<HfMuxState, RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn set_mux(
-        &mut self,
-        _: &RecvMessage,
-        _state: HfMuxState,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn get_dev(
-        &mut self,
-        _: &RecvMessage,
-    ) -> Result<HfDevSelect, RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn set_dev(
-        &mut self,
-        _: &RecvMessage,
-        _state: HfDevSelect,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn check_dev(
-        &mut self,
-        _: &RecvMessage,
-        _state: HfDevSelect,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn hash(
-        &mut self,
-        _: &RecvMessage,
-        _addr: u32,
-        _len: u32,
-    ) -> Result<[u8; SHA256_SZ], RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn get_persistent_data(
-        &mut self,
-        _: &RecvMessage,
-    ) -> Result<HfPersistentData, RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn write_persistent_data(
-        &mut self,
-        _: &RecvMessage,
-        _dev_select: HfDevSelect,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn hash_significant_bits(
-        &mut self,
-        _: &RecvMessage,
-        _dev: HfDevSelect,
-    ) -> Result<(), RequestError<HfError>> {
-        Err(self.err.into())
-    }
-
-    fn get_cached_hash(
-        &mut self,
-        _: &RecvMessage,
-        _dev: HfDevSelect,
-    ) -> Result<[u8; SHA256_SZ], RequestError<HfError>> {
-        Err(self.err.into())
-    }
-}
-
-impl NotificationHandler for FailServer {
-    fn current_notification_mask(&self) -> u32 {
-        0
-    }
-
-    fn handle_notification(&mut self, _bits: u32) {
-        unreachable!()
-    }
-}
-
 pub mod idl {
     use drv_hf_api::{
         HfDevSelect, HfError, HfMuxState, HfPersistentData, HfProtectMode,
