+++ drv/hf-fail-server/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "drv-hf-fail-server"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+idol-runtime = { workspace = true }
+num-traits = { workspace = true }
+zerocopy = { workspace = true }
+zerocopy-derive = { workspace = true }
+serde.workspace = true
+hubpack.workspace = true
+
+drv-hf-api = { path = "../hf-api" }
+drv-hash-api = { path = "../hash-api" }
+userlib = { path = "../../sys/userlib", features = ["panic-messages"] }
+
+[build-dependencies]
+build-util = {path = "../../build/util"}
+idol.workspace = true
+
+# None of these features do anything in the mock server; they exist only to
+# match the feature list in the real hf server (allowing a quick substitution
+# for real -> mock in app.toml without having to muck with features).
+[features]
+host_access = []
+hash = []
+h743 = []
+h753 = []
+
+[lib]
+test = false
+doctest = false
+bench = false
+
+[lints]
+workspace = true
